Changed all Thor enums to enum classes. Updated the Chimera spi interface.

+++ Thor/Thor/include/serial.hpp
@@ -151,7 +151,7 @@ namespace Thor
 
 				/* Specialized functions used only when interfacing with the Chimera HAL and are not intended be called by the user. */
 				#if defined(USING_CHIMERA)
-				Chimera::Serial::Status cbegin(uint32_t baud = Chimera::Serial::BaudRate::SERIAL_BAUD_115200,
+				Chimera::Serial::Status cbegin(uint32_t baud = static_cast<uint32_t>(Chimera::Serial::BaudRate::SERIAL_BAUD_115200),
 												Chimera::Serial::Modes tx_mode = Chimera::Serial::Modes::BLOCKING,
 												Chimera::Serial::Modes rx_mode = Chimera::Serial::Modes::BLOCKING);
